+++ index.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <script type="module" src="./index.js"></script>
+    <link rel="stylesheet" href="./Background/style.css">
+
 </head>
 <style>
     body{
@@ -44,12 +45,16 @@
 
         <div id="game-container">
             <div id="scrollBackground"></div>
-            <script src="./Background/scrollBackground.js"></script>
-            <link rel="stylesheet" href="./Background/style.css">
         </div>
 
         <!-- <div id="mainGame" class="mainGame"></div> -->
     </div>
     
 </body>
+<footer>
+
+    <script src="./Background/scrollBackground.js"></script>
+    <script type="module" src="./playerMovement.js"></script>
+    <script type="module" src="./index.js"></script>
+</footer>
 </html>
